requirements.txt
# Core numerical & system
numpy==1.26.4
scipy==1.14.1
psutil==6.0.0

# Audio I/O
sounddevice==0.5.1
soundfile==0.13.1

pyttsx3==2.98

# PyTorch stack (LOCKED)
torch==2.2.2
torchaudio==2.2.2

# Hugging Face core
transformers==4.46.1
tokenizers==0.20.3
huggingface-hub==0.34.4
sentencepiece==0.2.0

# ASR
openai-whisper==20240930
jiwer==3.1.0

# Translation eval
sacrebleu==2.5.1
tqdm==4.67.1

# IndicConformer dependency (CRITICAL)
onnxruntime==1.21.1

# TTS
git+https://github.com/huggingface/parler-tts.git@d108732cd57788ec86bc857d99a6cabd66663d68
